+++ src/shared/state.ts
@@ -20,6 +20,7 @@ export function getDefaultState(): AppState {
       zoomFactor: 1,
       activeTheme: 'system',
       minimizeToTray: false,
+      disableFontLigatures: false,
     },
     logins: [],
   }